"This series fixes some more errors found by clang, with the aim of enabling
clang/zapcc as a supported compiler.  A single issue remains, but it's
probably in std::experimental::optional::swap(); not in our code."

* tag 'clang/2/v1' of https://github.com/avikivity/scylla:
  sstable_test: avoid passing negative non-type template arguments to unsigned parameters
  UUID: add more comparison operators
  sstable_datafile_test: avoid string_view user-defined literal conversion operator
  mutation_source_test: avoid template function without template keyword
  cql_query_test: define static variable
  cql_query_test: add braces for single-item collection initializers
  storage_service: don't use typeid(temporary)
  logalloc: remove unused max_occupancy_for_compaction
  storage_proxy: drop overzealous use of __int128_t in recently-modified-no-read-repair logic
  storage_proxy: drop unused member access from return value
  storage_proxy: fix reference bound to temporary in data_read_resolver::less_compare
  read_repair_decision: fix operator<<(std::ostream&, ...)

+++ tests/sstable_test.cc
@@ -170,7 +170,7 @@ SEASTAR_TEST_CASE(small_summary_query_fail) {
 }
 
 SEASTAR_TEST_CASE(small_summary_query_negative_fail) {
-    return summary_query_fail<-2, 0, 5>(uncompressed_schema(), "tests/sstables/uncompressed", 1);
+    return summary_query_fail<-uint64_t(2), 0, 5>(uncompressed_schema(), "tests/sstables/uncompressed", 1);
 }
 
 SEASTAR_TEST_CASE(big_summary_query_0) {
@@ -193,7 +193,7 @@ SEASTAR_TEST_CASE(missing_summary_query_fail) {
 }
 
 SEASTAR_TEST_CASE(missing_summary_query_negative_fail) {
-    return summary_query_fail<-2, 0, 5>(uncompressed_schema(), "tests/sstables/uncompressed", 2);
+    return summary_query_fail<-uint64_t(2), 0, 5>(uncompressed_schema(), "tests/sstables/uncompressed", 2);
 }
 
 SEASTAR_TEST_CASE(missing_summary_interval_1_query_ok) {
